requirements.txt
streamlit==1.35.0
python-dotenv==1.1.1
openai==1.60.0
langchain-openai==0.3.5
